PAlib 0.77a

+++ PAlib/lib/arm9/PA/PA_Sound.c
@@ -1,17 +1,16 @@
 #include "PA9.h"
 
 PA_SoundOptions PA_SoundOption; // Default sound formats...
-u8 *FS_mod;
-u8 *FS_wav[16];
+u32 *FS_mod;
+u32 *FS_wav[16];
 
 void PA_SetDefaultSound(u8 volume, s16 freq, s16 format){
 	PA_SoundOption.volume = volume;
 	PA_SoundOption.freq = freq;
 	PA_SoundOption.format = format;
 }
-
-		 
-void PA_PlaySoundEx(u8 PA_Channel, const void* data, u32 length, u8 volume, s16 freq, s16 format){
+	 
+void PA_PlaySoundEx(u8 PA_Channel, const void* data, s32 length, u8 volume, s16 freq, s16 format){
 //		PA_Channel = 15 - PA_Channel;
         snd.data[PA_Channel].data = data;
         snd.data[PA_Channel].len = length;
